refactor FixedHeader and JobList into subcomponents
